+++ src/passes/circuit_opt_passes.cpp
@@ -5,6 +5,7 @@
 
 namespace qristal {
 
+/// Optimization_pass class
 /// Constructor
 optimization_pass::optimization_pass(const std::string &name)
     : m_plugin_name(name) {
@@ -14,6 +15,7 @@ optimization_pass::optimization_pass(const std::string &name)
         "' is not available. Please check your input or the installation.");
   }
 }
+
 /// Returns name of the pass
 std::string optimization_pass::get_name() const { return m_plugin_name; }
 
@@ -24,6 +26,24 @@ std::string optimization_pass::get_description() const {
 
 /// Runs the pass over the circuit IR node
 void optimization_pass::apply(CircuitBuilder &circuit) {
-  xacc::getIRTransformation(m_plugin_name)->apply(circuit.get(), nullptr);
+    xacc::getIRTransformation(m_plugin_name)->apply(circuit.get(), nullptr); 
+}
+
+/// sequence_pass class
+/// Constructor
+sequence_pass::sequence_pass(const std::vector<std::string> &pass_list): m_pass_list(pass_list) {}
+/// Returns name of the pass
+std::string sequence_pass::get_name() const { return "sequence_pass"; }
+
+/// Returns the pass description
+std::string sequence_pass::get_description() const {
+  return "A transformation that runs a sequence of optimization passes.";
+}
+
+/// Runs the pass over the circuit IR node
+void sequence_pass::apply(CircuitBuilder &circuit) {
+  for (const auto& s : m_pass_list) {
+    xacc::getIRTransformation(s)->apply(circuit.get(), nullptr);
+  }
 }
 }