+++ CHANGELOG.md
@@ -26,6 +26,8 @@ Qristal is a full-stack SDK for quantum accelerators.
 - Added docs to result-related private member variables of `session.hpp`
 - Build docs in CI with `-DINSTALL_MISSING=ON`
 - Added integration tests for circuit shot parallelisation via MPI
+- Added sequence_pass class to support applying multiple optimization passes in sequence
+- Added Python example demonstrating the use of sequence_pass
 
 ### Fixed